Move game config to atom
